Edwin/db105_linebot/requirements.txt
Flask
Flask-Bootstrap
Flask-Bootstrap4
Flask-Moment
Flask-WTF
flask_paginate
WTForms
Jinja2
line-bot-sdk
redis
confluent_kafka
elasticsearch
gunicorn